abstracted arrow logic to another component TimelineArrow + did some initial styling on arrow
