+++ frontend/scss/timeline.scss
@@ -7,9 +7,35 @@
 .showcase-item {
     cursor: pointer;
     margin-bottom: 16px;
-    
+
     img {
         width: 15vw;
         height: 50vh;
     }
+}
+
+.arrow {
+    font-size: 5rem;
+    background: none;
+    //border: none;
+}
+
+.arrow:hover {
+    padding-left: 1px;
+}
+
+.date {
+    font-size: 1rem;
+    color: lightgray;
+}
+
+.date:hover {
+    color: darkblue;
+    padding: 0px;
+}
+
+.col {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 }
